Create Quantcast-Top-Websites-By-US-Visitors.txt
This list ranks websites based on the number of people in the United States who visit each site within a month. It includes sites with estimated traffic as well as sites with traffic that Quantcast has independently verified. Quantcast directly measures site traffic through the implementation of the Quantcast asynchronous tag on each web site.